Putting PatchEmbedOmnivore into omnivore model and resolves some comments
