+++ torchmultimodal/models/omnivore.py
@@ -1,8 +1,10 @@
+from typing import Callable, Optional, Tuple
+
 from torch import nn
 from torchmultimodal.architectures.omnivore import OmnivoreArchitecture
 from torchmultimodal.modules.encoders.swin_transformer_3d_encoder import (
+    PatchEmbed3d,
     SwinTransformer3dEncoder,
-    PatchEmbedOmnivore,
 )
 
 
@@ -31,22 +33,64 @@ def _multimodal_head(input_dim: int) -> nn.ModuleDict:
     )
 
 
+class PatchEmbedOmnivore(nn.Module):
+    """Patch Embedding strategy for Omnivore model
+    It will use common PatchEmbed3d for image and video,
+    for single view depth image it will have separate embedding for the depth channel
+    and add the embedding result with the RGB channel
+    reference: https://arxiv.org/abs/2201.08377
+
+    Args:
+        patch_size (Tuple[int, int, int]): Patch token size. Default: (2, 4, 4)
+        embed_dim (int): Number of linear projection output channels. Default: 96
+        norm_layer (nn.Module, optional): Normalization layer. Default: None
+    """
+
+    def __init__(
+        self,
+        patch_size: Tuple[int, int, int] = (2, 4, 4),
+        embed_dim: int = 96,
+        norm_layer: Optional[Callable[..., nn.Module]] = None,
+    ):
+        super().__init__()
+        self.patch_embed = PatchEmbed3d(
+            patch_size=patch_size,
+            embed_dim=embed_dim,
+            norm_layer=norm_layer,
+        )
+
+        self.depth_patch_embed = PatchEmbed3d(
+            patch_size=patch_size,
+            in_channels=1,
+            embed_dim=embed_dim,
+            norm_layer=norm_layer,
+        )
+
+    def forward(self, x):
+        # x: B C D H W
+        # Note: D here represent time
+        assert x.ndim == 5
+        has_depth = x.shape[1] == 4
+
+        if has_depth:
+            x_rgb = self.patch_embed(x[:, :3, ...])
+            x_d = self.depth_patch_embed(x[:, 3:, ...])
+            x = x_rgb + x_d
+        else:
+            x = self.patch_embed(x)
+        return x
+
+
 def omnivore_swin_t() -> nn.Module:
-    embed_dim = 96
-    norm_layer = nn.LayerNorm
-    patch_embed = PatchEmbedOmnivore(
-        patch_size=(2, 4, 4),
-        embed_dim=embed_dim,
-        norm_layer=norm_layer,
-    )
     encoder = SwinTransformer3dEncoder(
-        embed_dim=embed_dim,
+        patch_size=(2, 4, 4),
+        embed_dim=96,
         depths=[2, 2, 6, 2],
         num_heads=[3, 6, 12, 24],
         window_size=(8, 7, 7),
         stochastic_depth_prob=0.2,
-        norm_layer=norm_layer,
-        patch_embed=patch_embed,
+        norm_layer=nn.LayerNorm,
+        patch_embed=PatchEmbedOmnivore,
     )
     heads = _multimodal_head(input_dim=encoder.num_features)
     return OmnivoreArchitecture(encoder, heads)
